requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
requests==2.32.3
python-dotenv==1.0.1
google-cloud-storage
beautifulsoup4==4.12.3
lxml==5.3.0
pdfplumber==0.11.4
